Move MatrixSet to separate file
