+++ sympy/matrices/expressions/matexpr.py
@@ -1,5 +1,5 @@
 from typing import Any, Callable
-from sympy.core.logic import FuzzyBool, fuzzy_and
+from sympy.core.logic import FuzzyBool
 
 from functools import wraps, reduce
 import collections
@@ -15,7 +15,6 @@ from sympy.simplify import simplify
 from sympy.utilities.misc import filldedent
 from sympy.assumptions.ask import ask, Q
 from sympy.multipledispatch import dispatch
-from sympy.sets.sets import Set
 
 
 def _sympifyit(arg, retval=None):
@@ -1269,63 +1268,6 @@ def _make_matrix(x):
         return x
     return ImmutableDenseMatrix([[x]])
 
-
-class MatrixSet(Set):
-    """
-    MatrixSet represents the set of matrices with ``shape = (n, m)`` over the
-    given set.
-
-    Examples
-    ========
-
-    >>> from sympy.matrices import MatrixSet, Matrix
-    >>> from sympy import S, I
-    >>> M = MatrixSet(2, 2, set=S.Reals)
-    >>> X = Matrix([[1, 2], [3, 4]])
-    >>> X in M
-    True
-    >>> X = Matrix([[1, 2], [I, 4]])
-    >>> X in M
-    False
-
-    """
-    is_empty = False
-
-    def __new__(cls, n, m, set):
-        n, m, set = _sympify(n), _sympify(m), _sympify(set)
-        cls._check_dim(n)
-        cls._check_dim(m)
-        if not isinstance(set, Set):
-            raise TypeError("{} should be an instance of Set.".format(set))
-        return Set.__new__(cls, n, m, set)
-
-    @property
-    def shape(self):
-        return self.args[:2]
-
-    @property
-    def set(self):
-        return self.args[2]
-
-    def _contains(self, other):
-        if not isinstance(other, MatrixExpr):
-            raise TypeError("{} should be an instance of MatrixExpr.".format(other))
-        if other.shape != self.shape:
-            are_symbolic = any(_sympify(x).is_Symbol for x in other.shape + self.shape)
-            if are_symbolic:
-                return None
-            return False
-        return fuzzy_and(self.set.contains(x) for x in other)
-
-    @classmethod
-    def _check_dim(cls, dim):
-        """Helper function to check invalid matrix dimensions"""
-        from sympy.core.assumptions import check_assumptions
-        ok = check_assumptions(dim, integer=True, nonnegative=True)
-        if ok is False:
-            raise ValueError(
-                "The dimension specification {} should be "
-                "a nonnegative integer.".format(dim))
 
 from .matmul import MatMul
 from .matadd import MatAdd
